New branch from 'create crud for category product ani

+++ src/main/java/am/itspace/sweetbakerystorecommon/service/CategoryService.java
@@ -1,12 +1,16 @@
 package am.itspace.sweetbakerystorecommon.service;
 
+import am.itspace.sweetbakerystorecommon.dto.categoryDto.CategoryResponseDto;
 import am.itspace.sweetbakerystorecommon.entity.Category;
 import am.itspace.sweetbakerystorecommon.entity.User;
 import am.itspace.sweetbakerystorecommon.repository.CategoryRepository;
+import am.itspace.sweetbakerystorecommon.security.CurrentUser;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,8 +23,8 @@ public class CategoryService {
     private final CategoryRepository categoryRepository;
 
     public Page<Category> findPaginated(Pageable pageable) {
-
-        return categoryRepository.findAll(pageable);
+        Page<Category> categoryPages = categoryRepository.findAll(pageable);
+        return new PageImpl<>(categoryPages.getContent(), pageable, categoryPages.getSize());
     }
 
     public Optional<Category> findByName(String name) {
@@ -28,8 +32,12 @@ public class CategoryService {
 
     }
 
-    public void deleteById(int id) {
-        categoryRepository.deleteById(id);
+    public void deleteById(int id, CurrentUser currentUser) {
+        Optional<Category> byId = categoryRepository.findById(id);
+        if (byId.isPresent() && currentUser.getUser().getId() == byId.get().getUser().getId()) {
+            categoryRepository.deleteById(id);
+        }
+        ResponseEntity.notFound().build();
     }
 
     public void save(Category category, User user) throws Exception {
@@ -48,4 +56,23 @@ public class CategoryService {
     public Long getCountOfCategories() {
         return categoryRepository.count();
     }
+
+    public Category findByCategoryId(int id) {
+        Optional<Category> byId = categoryRepository.findById(id);
+        return byId.orElse(null);
+    }
+
+
+    public Category saveCategory(Category category) {
+        if(category==null){
+            throw new RuntimeException("Category can't be null!");
+        }
+        return categoryRepository.save(category);
+    }
+
+    public Optional<Category> findById(int id) {
+        return categoryRepository.findById(id);
+    }
+
+
 }
